+++ core/unused/RecalculateStats_OLD2.lua
@@ -1,0 +1,95 @@
+-- RecalculateStats
+
+if true then return end
+
+-- Temporary flawed implementation that simply post-hooks
+-- like before while we wait for a better implementation by sarn
+-- (Re-adding all the mid-hooks is time consuming)
+
+--[[
+**[!]** The functionality of this class is subject to change in the future once a more accurate system is established.
+]]
+
+RecalculateStats = new_class()
+
+run_once(function()
+    __recalcstats_cache = CallbackCache.new()
+end)
+
+
+
+-- ========== Static Methods ==========
+
+--@section Static Methods
+
+--@static
+--@return       number
+--@param        fn          | function  | The function to register. <br>The parameter for it is `actor`.
+--@overload
+--@return       number
+--@param        priority    | number    | The priority of the function. <br>Higher values run before lower ones; can be negative. <br>`Callback.Priority.NORMAL` (`0`) by default.
+--@param        fn          | function  | The function to register. <br>The parameter for it is `actor`.
+--[[
+Registers a function for stat recalculation.
+Returns the unique ID of the registered function.
+
+*Technical:* This function will run in `recalculate_stats` post-hook.
+
+**Priority Convention**
+To allow for a decent amount of space between priorities,
+use the enum values in @link {`Callback.Priority` | Callback#Priority}.
+If you need to be more specific than that, try to keep a distance of at least `100`.
+]]
+RecalculateStats.add = function(NAMESPACE, arg1, arg2)
+    if type(arg1) == "function" then
+        return __recalcstats_cache:add(arg1, NAMESPACE)
+    end
+    return __recalcstats_cache:add(arg2, NAMESPACE, arg1)
+end
+
+
+--@static
+--@return       function
+--@param        id          | number    | The unique ID of the registered function to remove.
+--[[
+Removes and returns a registered stat recalculation function.
+The ID is the one from @link {`RecalculateStats.add` | RecalculateStats#add}.
+]]
+RecalculateStats.remove = function(id)
+    return __recalcstats_cache:remove(id)
+end
+
+
+--@static
+--[[
+Removes all registered stat recalculation functions from your namespace.
+
+Automatically called when you hotload your mod.
+]]
+RecalculateStats.remove_all = function(NAMESPACE)
+    __recalcstats_cache:remove_all(NAMESPACE)
+end
+table.insert(_clear_namespace_functions, RecalculateStats.remove_all)
+
+
+
+-- ========== Hooks ==========
+
+gm.post_script_hook(gm.constants.recalculate_stats, function(self, other, result, args)
+    local actor = Instance.wrap(self)
+
+    -- Call registered functions with wrapped arg
+    __recalcstats_cache:loop_and_call_functions(function(fn_table)
+        local status, err = pcall(fn_table.fn, actor)
+        if not status then
+            if (err == nil)
+            or (err == "C++ exception") then err = "GameMaker error (see above)" end
+            log.warning("\n"..fn_table.namespace..": RecalculateStats (ID '"..fn_table.id.."') failed to execute fully.\n"..err)
+        end
+    end)
+end)
+
+
+
+-- Public export
+__class.RecalculateStats = RecalculateStats